src/test/java/http/RequestHeaderTest.java
package http;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.io.*;

import static org.assertj.core.api.Assertions.assertThat;

class RequestHeaderTest {
    @Test
    @DisplayName("헤더를 잘 파싱하는지 테스트")
    void headerParseTest() throws IOException {
        String request = "Host: localhost:8080\n" +
                "Connection: keep-alive\n" +
                "Content-Length: 93\n" +
                "Content-Type: application/x-www-form-urlencoded\n" +
                "Accept: */*\n" +
                "\n" +
                "userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net";

        InputStream in = new ByteArrayInputStream(request.getBytes());
        BufferedReader br = new BufferedReader(new InputStreamReader(in));

        RequestHeader requestHeader = new RequestHeader(br);

        assertThat(requestHeader.getContentLength()).isEqualTo(93);
    }
}